requirements.txt
aniso8601==9.0.1
attrs==23.2.0
better-exceptions==0.3.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-restx==1.3.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
Mopidy==3.4.2
mysql-connector-python==8.4.0
pretty-errors==1.2.25
Pygments==2.18.0
pykka==4.0.2
python-mpv==1.0.6
pytz==2024.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
sql==2022.4.0
tornado==6.4
urllib3==2.2.1
Werkzeug==3.0.3
